# public/swagger.yaml
openapi: 3.0.0
info:
  description: API REST para facturación electrónica SUNAT (bendey-api).
  version: "1.3"
  title: bendey-api
  contact:
    name: Giancarlos Salas
    email: giansalex@gmail.com
...
security:
  - ApiToken: []
paths:
  /invoice/send:
    post:
      tags:
        - invoice
      summary: Enviar una Factura o Boleta
      description: Retorna la respuesta de SUNAT y el XML firmado
      operationId: sendInvoice
      requestBody:
        $ref: "#/components/requestBodies/Invoice"
      responses:
        "200":
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "400":
          description: Error de Validación
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
  /invoice/xml:
    post:
      tags:
        - invoice
      summary: Generar el XML de una factura o boleta
      description: Retorna el XML
      operationId: getInvoiceXml
      requestBody:
        $ref: "#/components/requestBodies/Invoice"
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Error de Validación
          content:
            text/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
  /invoice/pdf:
    post:
      tags:
        - invoice
      summary: Generar el PDF de una factura o boleta
      description: Retorna el Pdf
      operationId: getInvoicePdf
      requestBody:
        $ref: "#/components/requestBodies/Invoice"
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Error de Validación
          content:
            application/pdf:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
...
  /despatch/send:
    post:
      tags:
        - despatch
      summary: Enviar una Guía de Remisión
      description: Retorna la respuesta de SUNAT y el XML firmado
      operationId: sendDespatch
      requestBody:
        $ref: "#/components/requestBodies/Despatch"
      responses:
        "200":
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummaryResponse"
        "400":
          description: Error de Validación
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
  /despatch/xml:
    post:
      tags:
        - despatch
      summary: Generar el XML de una Guía de Remisión
      description: Retorna el XML
      operationId: getDespatchXml
      requestBody:
        $ref: "#/components/requestBodies/Despatch"
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Error de Validación
          content:
            text/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
  /despatch/status:
    get:
      tags:
        - despatch
      summary: Obtiene el estado del ticket
      description: Retorna el CDR
      operationId: getDespatchStatus
      parameters:
        - name: ticket
          in: query
          description: Ticket del comprobante
          schema:
            type: string
        - name: ruc
          in: query
          description: RUC de empresa (Multiples Empresas)
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResult"
  /despatch/pdf:
    post:
      tags:
        - despatch
      summary: Generar el Pdf de una Guía de Remisión
      description: Retorna el Pdf
      operationId: getDespatchPdf
      requestBody:
        $ref: "#/components/requestBodies/Despatch"
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Error de Validación
          content:
            application/Pdf:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
...
  /reversion/send:
    post:
      tags:
        - reversion
      summary: Enviar un Resumen de Reversiones
      description: Retorna la respuesta de SUNAT y el XML firmado
      operationId: sendReversion
      requestBody:
        $ref: "#/components/requestBodies/Reversion"
      responses:
        "200":
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummaryResponse"
        "400":
          description: Error de Validación
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
  /reversion/xml:
    post:
      tags:
        - voided
      summary: Generar el XML de un Resumen de Reversiones
      description: Retorna el XML
      operationId: getReversionXml
      requestBody:
        $ref: "#/components/requestBodies/Reversion"
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Error de Validación
          content:
            text/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
  /reversion/pdf:
    post:
      tags:
        - voided
      summary: Generar el pdf de un Resumen de Reversiones
      description: Retorna el pdf
      operationId: getReversionpdf
      requestBody:
        $ref: "#/components/requestBodies/Reversion"
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Error de Validación
          content:
            application/pdf:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
  /reversion/status:
    get:
      tags:
        - voided
      summary: Obtiene el estado del ticket
      description: Retorna el XML
      operationId: getReversionStatus
      parameters:
        - name: ticket
          in: query
          description: Ticket de Resumen de Reversiones
          schema:
            type: string
        - name: ruc
          in: query
          description: RUC de empresa (Multiples Empresas)
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResult"
...
servers:
  - url: http://lycet.api/api/v1
  - url: https://lycet.api/api/v1
components:
  requestBodies:
...
    Invoice:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Invoice"
      description: Comprobante
      required: true
...
    Despatch:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Despatch"
      description: Comprobante
      required: true
...
    Reversion:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Reversion"
      description: Comprobante
      required: true
  securitySchemes:
    ApiToken:
      name: token
      type: apiKey
      in: query
      description: Api Token
  schemas:
    Address:
      type: object
      properties:
        ubigueo:
          type: string
        codigoPais:
          type: string
        departamento:
          type: string
        provincia:
          type: string
        distrito:
          type: string
        urbanizacion:
          type: string
        direccion:
          type: string
        codLocal:
          type: string
    Client:
      type: object
      properties:
        tipoDoc:
          type: string
        numDoc:
          type: string
        rznSocial:
          type: string
        address:
          $ref: "#/components/schemas/Address"
        email:
          type: string
        telephone:
          type: string
    Company:
      type: object
      properties:
        ruc:
          type: string
        razonSocial:
          type: string
        nombreComercial:
          type: string
        address:
          $ref: "#/components/schemas/Address"
        email:
          type: string
        telephone:
          type: string
    Charge:
      type: object
      properties:
        codTipo:
          type: string
        factor:
          type: string
        monto:
          type: number
          format: float
        montoBase:
          type: number
          format: float
    DetailAttribute:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        value:
          type: string
        fecInicio:
          type: string
          format: date-time
        fecFin:
          type: string
          format: date-time
        duracion:
          type: integer
    SaleDetail:
      type: object
      properties:
        unidad:
          type: string
        cantidad:
          type: number
          format: float
        codProducto:
          type: string
        codProdSunat:
          type: string
        codProdGS1:
          type: string
        descripcion:
          type: string
        mtoValorUnitario:
          type: number
          format: float
        descuento:
          type: number
          format: float
        igv:
          type: number
          format: float
        tipAfeIgv:
          type: string
        isc:
          type: number
          format: float
        tipSisIsc:
          type: string
        totalImpuestos:
          type: number
          format: float
        mtoPrecioUnitario:
          type: number
          format: float
        mtoValorVenta:
          type: number
          format: float
        mtoValorGratuito:
          type: number
          format: float
        mtoBaseIgv:
          type: number
          format: float
        porcentajeIgv:
          type: number
          format: float
        mtoBaseIsc:
          type: number
          format: float
        porcentajeIsc:
          type: number
          format: float
        mtoBaseOth:
          type: number
          format: float
        porcentajeOth:
          type: number
          format: float
        otroTributo:
          type: number
          format: float
        icbper:
          type: number
          format: float
        factorIcbper:
          type: number
          format: float
        cargos:
          type: array
          items:
            $ref: "#/components/schemas/Charge"
        descuentos:
          type: array
          items:
            $ref: "#/components/schemas/Charge"
        atributos:
          type: array
          items:
            $ref: "#/components/schemas/DetailAttribute"
    Legend:
      type: object
      properties:
        code:
          type: string
        value:
          type: string
    Document:
      type: object
      properties:
        tipoDoc:
          type: string
        nroDoc:
          type: string
    SalePerception:
      type: object
      properties:
        codReg:
          type: string
        porcentaje:
          type: number
          format: float
        mtoBase:
          type: number
          format: float
        mto:
          type: number
          format: float
        mtoTotal:
          type: number
          format: float
    Direction:
      type: object
      properties:
        ubigueo:
          type: string
        direccion:
          type: string
    EmbededDespatch:
      type: object
      properties:
        llegada:
          $ref: "#/components/schemas/Direction"
        partida:
          $ref: "#/components/schemas/Direction"
        transportista:
          $ref: "#/components/schemas/Client"
        nroLicencia:
          type: string
        transpPlaca:
          type: string
        transpCodeAuth:
          type: string
        transpMarca:
          type: string
        modTraslado:
          type: string
        pesoBruto:
          type: number
          format: float
        undPesoBruto:
          type: string
    Prepayment:
      type: object
      properties:
        tipoDocRel:
          type: string
        nroDocRel:
          type: string
        total:
          type: number
          format: float
    Detraction:
      type: object
      properties:
        percent:
          type: number
          format: float
        mount:
          type: number
          format: float
        ctaBanco:
          type: string
        codMedioPago:
          type: string
        codBienDetraccion:
          type: string
        valueRef:
          type: number
          format: float
    Invoice:
      type: object
      properties:
        tipoDoc:
          type: string
        serie:
          type: string
        correlativo:
          type: string
        fechaEmision:
          type: string
          format: date-time
        formaPago:
          $ref: "#/components/schemas/PaymentTerms"
        cuotas:
          type: array
          items:
            $ref: "#/components/schemas/Cuota"
        client:
          $ref: "#/components/schemas/Client"
        company:
          $ref: "#/components/schemas/Company"
        tipoMoneda:
          type: string
        sumOtrosCargos:
          type: number
          format: float
        mtoOperGravadas:
          type: number
          format: float
        mtoOperInafectas:
          type: number
          format: float
        mtoOperExoneradas:
          type: number
          format: float
        mtoOperExportacion:
          type: number
          format: float
        mtoIGV:
          type: number
          format: float
        mtoIGVGratuitas:
          type: number
          format: float
        mtoISC:
          type: number
          format: float
        mtoOtrosTributos:
          type: number
          format: float
        icbper:
          type: number
          format: float
        valorVenta:
          type: number
          format: float
        subTotal:
          type: number
          format: float
        mtoImpVenta:
          type: number
          format: float
        details:
          type: array
          items:
            $ref: "#/components/schemas/SaleDetail"
        legends:
          type: array
          items:
            $ref: "#/components/schemas/Legend"
        guias:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        relDocs:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        observacion:
          type: string
        compra:
          type: string
        mtoBaseIsc:
          type: number
          format: float
        mtoBaseOth:
          type: number
          format: float
        totalImpuestos:
          type: number
          format: float
        ublVersion:
          type: string
        tipoOperacion:
          type: string
        fecVencimiento:
          type: string
          format: date-time
        sumDsctoGlobal:
          type: number
          format: float
        mtoDescuentos:
          type: number
          format: float
        mtoOperGratuitas:
          type: number
          format: float
        totalAnticipos:
          type: number
          format: float
        perception:
          $ref: "#/components/schemas/SalePerception"
        guiaEmbebida:
          $ref: "#/components/schemas/EmbededDespatch"
        anticipos:
          type: array
          items:
            $ref: "#/components/schemas/Prepayment"
        detraccion:
          $ref: "#/components/schemas/Detraction"
        seller:
          $ref: "#/components/schemas/Client"
        direccionEntrega:
          $ref: "#/components/schemas/Address"
        descuentos:
          type: array
          items:
            $ref: "#/components/schemas/Charge"
        cargos:
          type: array
          items:
            $ref: "#/components/schemas/Charge"
        mtoCargos:
          type: number
          format: float
...
    PaymentTerms:
      type: object
      properties:
        moneda:
          type: string
        tipo:
          type: string
        monto:
          type: number
          format: float
    Cuota:
      type: object
      properties:
        moneda:
          type: string
        monto:
          type: number
          format: float
        fechaPago:
          type: string
          format: date-time
...
    VoidedDetail:
      type: object
      properties:
        tipoDoc:
          type: string
        serie:
          type: string
        correlativo:
          type: string
        desMotivoBaja:
          type: string
...
    Reversion:
      type: object
      properties:
        correlativo:
          type: string
        fecGeneracion:
          type: string
          format: date-time
        fecComunicacion:
          type: string
          format: date-time
        company:
          $ref: "#/components/schemas/Company"
        details:
          type: array
          items:
            $ref: "#/components/schemas/VoidedDetail"
...
    Puerto:
      type: object
      properties:
        codigo:
          type: string
        nombre:
          type: string
    Transportist:
      type: object
      properties:
        tipoDoc:
          type: string
        numDoc:
          type: string
        rznSocial:
          type: string
        nroMtc:
          type: string
        placa:
          type: string
        choferTipoDoc:
          type: string
        choferDoc:
          type: string
    Vehicle:
      type: object
      properties:
        placa:
          type: string
        nroCirculacion:
          type: string
        codEmisor:
          type: string
        nroAutorizacion:
          type: string
        secundarios:
          type: array
          items:
            $ref: '#/components/schemas/Vehicle'
    Driver:
      type: object
      properties:
        tipo:
          type: string
        tipoDoc:
          type: string
        nroDoc:
          type: string
        nombres:
          type: string
        apellidos:
          type: string
        licencia:
          type: string
    Shipment:
      type: object
      properties:
        codTraslado:
          type: string
        desTraslado:
          type: string
        sustentoPeso:
          type: string
        indTransbordo:
          type: boolean
        indicadores:
          type: array
          items:
            type: string
        pesoItems:
          type: number
          format: float
        pesoTotal:
          type: number
          format: float
        undPesoTotal:
          type: string
        numBultos:
          type: integer
        modTraslado:
          type: string
        fecTraslado:
          type: string
          format: date-time
        numContenedor:
          type: string
        contenedores:
          type: array
          items:
            type: string
        codPuerto:
          type: string
        puerto:
          $ref: '#/components/schemas/Puerto'
        aeropuerto:
          $ref: '#/components/schemas/Puerto'
        transportista:
          $ref: '#/components/schemas/Transportist'
        vehiculo:
          $ref: '#/components/schemas/Vehicle'
        choferes:
          type: array
          items:
            $ref: '#/components/schemas/Driver'
        llegada:
          $ref: '#/components/schemas/Direction'
        partida:
          $ref: '#/components/schemas/Direction'
    AdditionalDoc:
      type: object
      properties:
        tipoDesc:
          type: string
        tipo:
          type: string
        nro:
          type: string
        emisor:
          type: string
    DespatchDetail:
      type: object
      properties:
        codigo:
          type: string
        descripcion:
          type: string
        unidad:
          type: string
        cantidad:
          type: number
          format: float
        codProdSunat:
          type: string
        atributos:
          type: array
          items:
            $ref: '#/components/schemas/DetailAttribute'
    Despatch:
      type: object
      properties:
        version:
          type: string
        tipoDoc:
          type: string
        serie:
          type: string
        correlativo:
          type: string
        observacion:
          type: string
        fechaEmision:
          type: string
          format: date-time
        company:
          $ref: '#/components/schemas/Company'
        destinatario:
          $ref: '#/components/schemas/Client'
        tercero:
          $ref: '#/components/schemas/Client'
        comprador:
          $ref: '#/components/schemas/Client'
        envio:
          $ref: '#/components/schemas/Shipment'
        addDocs:
          type: array
          items:
            $ref: '#/components/schemas/AdditionalDoc'
        details:
          type: array
          items:
            $ref: '#/components/schemas/DespatchDetail'
        name:
          type: string
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    CdrResponse:
      type: object
      properties:
        accepted:
          type: boolean
          readOnly: true
        id:
          type: string
        code:
          type: string
        description:
          type: string
        notes:
          type: array
          items:
            type: string
    BillResult:
      type: object
      properties:
        success:
          type: boolean
        error:
          $ref: "#/components/schemas/Error"
        cdrZip:
          type: string
          format: byte
        cdrResponse:
          $ref: "#/components/schemas/CdrResponse"
    SummaryResult:
      type: object
      properties:
        success:
          type: boolean
        error:
          $ref: "#/components/schemas/Error"
        ticket:
          type: string
    StatusResult:
      type: object
      properties:
        cdrZip:
          type: string
          format: byte
        cdrResponse:
          $ref: "#/components/schemas/CdrResponse"
        success:
          type: boolean
        error:
          $ref: "#/components/schemas/Error"
        code:
          type: string
    DocumentResponse:
      type: object
      properties:
        xml:
          type: string
        hash:
          type: string
          description: Resumen de la Firma Digital
        sunatResponse:
          $ref: "#/components/schemas/BillResult"
    SummaryResponse:
      type: object
      properties:
        xml:
          type: string
        hash:
          type: string
          description: Resumen de la Firma Digital
        sunatResponse:
          $ref: "#/components/schemas/SummaryResult"
    ValidationResponse:
      type: object
      properties:
        message:
          type: string
        field:
          type: string
...